+++ include/RAJA/policy/cuda/shared_memory.hpp
@@ -59,7 +59,7 @@ struct SharedMemory<cuda_shmem, T, NumElem> {
   static constexpr size_t size = NumElem;
   static constexpr size_t num_bytes = NumElem*sizeof(T);
 
-  ptrdiff_t offset; // offset into dynamic shared memory, in bytes
+  int offset; // offset into dynamic shared memory, in bytes
   void *parent;     // pointer to original object
 
   RAJA_INLINE
